Oh God, I don't even know. Let's see if I can do this:
-Simplified and Abstracted Physics out of the Player class. It is now housed
 in PhysicsHandler.cs
 One note here is that "Short Hopping" isn't possible right now. For that, I need a way to determine how long
 the jump button was pressed. If someone has a better idea, have at it.
 Also, FALLING does not work, since isOnGround is not updated correctly in collision detection yet. 

-Player now uses the InputMonitor (Though not as a service) and uses the new Physics. In addition to that
 I've made a change to how it uses Animations. I've introduced a Role class which subclasses into specific roles.
 
For example, Player.cs now has the role of "Cowboy" which really only houses the animation data. This allows you to change
to another role, such as "Space Cowboy" and still use the same "states" with different animations.

This is intended to be used with NPCs and Enemies as well. So, Enemy, NPC, or whatever would be a subclass of Role, and say
TUMBLEWEED.cs would have an instance of that subclass. I don't think it's too complicated, but it was a pain in the ass to do.

I think that's it for now. Brett, I'll implement your interfaces on Player later today.